+++ packages/components/scripts/generateDSRComponents.ts
@@ -743,7 +743,12 @@ $&`
           .replace(/this\.props\.value = state;/, '')
           .replace(/formDisabledCallback\(disabled: boolean\)/, 'formDisabledCallback()')
           .replace(/formStateRestoreCallback\(state: string\)/, 'formStateRestoreCallback()');
-      } else if (tagName === 'p-input-number' || tagName === 'p-input-search' || tagName === 'p-input-text') {
+      } else if (
+        tagName === 'p-input-number' ||
+        tagName === 'p-input-search' ||
+        tagName === 'p-input-text' ||
+        tagName === 'p-input-email'
+      ) {
         newFileContent = newFileContent
           .replace(/@AttachInternals\(\)/, '')
           .replace(