+++ packages/storefront/src/utils/generator/generator.tsx
@@ -64,6 +64,8 @@ import {
   type PInputSearchProps,
   PInputText,
   type PInputTextProps,
+  PInputEmail,
+  type PInputEmailProps,
   PLink,
   type PLinkProps,
   PLinkPure,
@@ -282,6 +284,7 @@ export type PropTypeMapping = {
   'p-input-number': PInputNumberProps;
   'p-input-search': PInputSearchProps;
   'p-input-text': PInputTextProps;
+  'p-input-email': PInputEmailProps;
   'p-link': PLinkProps;
   'p-link-pure': PLinkPureProps;
   'p-link-social': PLinkSocialProps;
@@ -364,6 +367,7 @@ const componentMap: Record<ConfiguratorTagNames, React.ElementType> = {
   'p-input-number': PInputNumber,
   'p-input-search': PInputSearch,
   'p-input-text': PInputText,
+  'p-input-email': PInputEmail,
   'p-link': PLink,
   'p-link-pure': PLinkPure,
   'p-link-social': PLinkSocial,